requirements.txt
# Core AI and API dependencies
groq>=0.8.0
langchain>=0.1.0
langchain-groq>=0.1.0
python-dotenv==1.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
requests==2.31.0

# Web search tool
duckduckgo-search==4.1.1

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.13.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
cffi>=1.15.1

# Email validation
email-validator==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2